requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.36.0
diffusers==0.25.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
gradio==4.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
scipy==1.11.4
scikit-image==0.22.0
matplotlib==3.8.2
